re-order and re-write based on SMA integration from ha core
