+++ custom_components/goodweproxy/const.py
@@ -5,7 +5,7 @@ DOMAIN = "goodweproxy"
 DEFAULT_NAME = "GoodWe Inverter"
 CONF_SERIAL = "serial"
 
-KEY_COORDINATOR = "coordinator"
-KEY_DEVICE_INFO = "device_info"
+GOODWE_COORDINATOR = "coordinator"
+GOODWE_DEVICE_INFO = "device_info"
 
 PLATFORMS = [Platform.SENSOR]
